requirements.txt
# # Core
pydantic==2.12.4
python-dotenv==1.2.1
PyYAML==6.0.3
joblib==1.5.2

# ML
numpy==2.3.5
pandas==2.3.3
scikit-learn==1.7.2
lightgbm==4.6.0
mlforecast==1.0.2
psycopg2-binary==2.9.11
mlflow==3.6.0
kaggle==1.8.0

# Testing
pytest==9.0.1

# -e . Commented just for creating docker image for airflow
